Update NavMakeAccount.java
The user can now actually create an account. This happens by making a txt file that has all the contents that have been inputted except the password, which defaults to "password". The username for the patient will be generated here and be comprised of the first letter of the user's first name followed by his/her last name which is then followed by the last two digits of the year he/she was born in. If the user name generated already exists, nothing will occur and the message will appear in the console.